Sample random covariance/concentration graph matrices.

Samples the covariance/concentration matrix corresponding to an undirected graph by either partial orthogonalization or diagonal dominance.

port(N = 1, ug = NULL, p = 5, d = 0.25, rentries = runif,
  zapzeros = TRUE)

diagdom(N = 1, ug = NULL, p = 5, d = 0.25, rentries = runif,
  k = NULL)

Arguments

N
sample size

ug
the undirected graph

p
Matrix dimension (number of nodes in ug if provided)

d
number in [0,1], the proportion of non-zero entries (if ug is not provided)

rentries
the random number generator for the non-zero entries (defaults to runif)

zapzeros
convert to zero extremely low entries

k
real number greater than 1, the desired condition number of the matrices in the resulting sample

Value

A three-dimensional array of length p*p*N

Details

Function port uses the method described in Córdoba et al. (2018). In summary, it consists on generating a random matrix Q and performing row-wise orthogonalization such that if i and j are not adjacent in ug, then the rows corresponding to such indices are orthogonalized, without violating previous orthogonalizations and without introducing unwanted independences. The resulting matrix after the process has finished is the cross product of Q.

We also provide an implementation of the most commonly used in the literature diagdom. By contrast, this method produces a random matrix M with zeros corresponding to missing edges in ug, and then enforces a dominant diagonal to ensure positive definiteness. Matrices produced by diagdom usually are better conditioned than those by port; however, they typically suffer from small off-diagonal entries, which can compromise model validation in Gaussian graphical models. This is avoided by port.

References

Córdoba, I., Varando, G., Bielza, C. and Larrañaga, P. A partial orthogonalization method for simulation covariance and concentration graph matrices. Proceedings of Machine Learning Research (PGM 2018), vol. 72, pp. 61 - 72, 2018.

Examples

# Generate a random undirected graph structure
ug <- igraph::sample_gnp(n = 3, p = 0.25)

# Generate 10 matrices complying with such random structure via 
# partial orthogonalization
gmat::port(N = 10, ug = ug)

#> , , 1
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 0.4193935 0.0000000 0.7623402
#> [2,] 0.0000000 0.0199438 0.0000000
#> [3,] 0.7623402 0.0000000 1.5760805
#> 
#> , , 2
#> 
#>          [,1]      [,2]     [,3]
#> [1,] 1.055225 0.0000000 1.092368
#> [2,] 0.000000 0.8495949 0.000000
#> [3,] 1.092368 0.0000000 1.130934
#> 
#> , , 3
#> 
#>          [,1]      [,2]     [,3]
#> [1,] 1.806167 0.0000000 1.289307
#> [2,] 0.000000 0.0246835 0.000000
#> [3,] 1.289307 0.0000000 1.275753
#> 
#> , , 4
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 1.2210752 0.0000000 0.5056447
#> [2,] 0.0000000 0.9897243 0.0000000
#> [3,] 0.5056447 0.0000000 0.3290136
#> 
#> , , 5
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 0.1923741 0.0000000 0.2601706
#> [2,] 0.0000000 0.9523532 0.0000000
#> [3,] 0.2601706 0.0000000 0.3647553
#> 
#> , , 6
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 0.6114823 0.0000000 0.6128217
#> [2,] 0.0000000 0.2432093 0.0000000
#> [3,] 0.6128217 0.0000000 1.0140311
#> 
#> , , 7
#> 
#>           [,1]      [,2]     [,3]
#> [1,] 0.9810017 0.0000000 1.049176
#> [2,] 0.0000000 0.3243603 0.000000
#> [3,] 1.0491759 0.0000000 1.175711
#> 
#> , , 8
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 0.1702705 0.0000000 0.2790246
#> [2,] 0.0000000 0.2163994 0.0000000
#> [3,] 0.2790246 0.0000000 0.5942012
#> 
#> , , 9
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 1.1092923 0.0000000 0.5371963
#> [2,] 0.0000000 0.2288734 0.0000000
#> [3,] 0.5371963 0.0000000 0.3197644
#> 
#> , , 10
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 0.2951385 0.0000000 0.6705835
#> [2,] 0.0000000 0.4870976 0.0000000
#> [3,] 0.6705835 0.0000000 1.5248435
#> 

# Generate 10 matrices complying with such random structure via 
# diagonal dominance 
gmat::diagdom(N = 10, ug = ug)

#> , , 1
#> 
#>           [,1]    [,2]      [,3]
#> [1,] 1.0794329 0.00000 0.2404612
#> [2,] 0.0000000 1.69108 0.0000000
#> [3,] 0.2404612 0.00000 0.7798505
#> 
#> , , 2
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 1.6300035 0.0000000 0.6388837
#> [2,] 0.0000000 0.6917072 0.0000000
#> [3,] 0.6388837 0.0000000 1.6154610
#> 
#> , , 3
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 1.2088045 0.0000000 0.5033951
#> [2,] 0.0000000 0.1346397 0.0000000
#> [3,] 0.5033951 0.0000000 0.6055200
#> 
#> , , 4
#> 
#>           [,1]       [,2]      [,3]
#> [1,] 0.7479930 0.00000000 0.2505233
#> [2,] 0.0000000 0.05403891 0.0000000
#> [3,] 0.2505233 0.00000000 0.2554769
#> 
#> , , 5
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 2.8964896 0.0000000 0.9790338
#> [2,] 0.0000000 0.2360891 0.0000000
#> [3,] 0.9790338 0.0000000 1.9707031
#> 
#> , , 6
#> 
#>           [,1]       [,2]      [,3]
#> [1,] 1.3764998 0.00000000 0.6565796
#> [2,] 0.0000000 0.02739074 0.0000000
#> [3,] 0.6565796 0.00000000 1.1441621
#> 
#> , , 7
#> 
#>           [,1]     [,2]      [,3]
#> [1,] 0.4878214 0.000000 0.0898489
#> [2,] 0.0000000 1.471459 0.0000000
#> [3,] 0.0898489 0.000000 0.6013474
#> 
#> , , 8
#> 
#>           [,1]     [,2]      [,3]
#> [1,] 2.0831136 0.000000 0.6305429
#> [2,] 0.0000000 1.847922 0.0000000
#> [3,] 0.6305429 0.000000 1.7351805
#> 
#> , , 9
#> 
#>           [,1]     [,2]      [,3]
#> [1,] 1.2130326 0.000000 0.8767677
#> [2,] 0.0000000 1.964381 0.0000000
#> [3,] 0.8767677 0.000000 1.3799229
#> 
#> , , 10
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 1.3832442 0.0000000 0.9302815
#> [2,] 0.0000000 0.8105404 0.0000000
#> [3,] 0.9302815 0.0000000 2.1871819
#>